Split EthereumTransaction and EthereumSignedTransaction
This is an attempt to bring things a bit closer to web3.js. Technically
unsigned transactions do not require many of the fields before being
sent to the provider node for signing. When signing locally or
representing a signed transaction however, there are much stricter
rules. EthereumTransaction represents the pending transaction object,
and EthereumSignedTransaction represents the immutable signed version
of the transaction. This also adds the ability to represent Contract
creation transactions which do not have a to address.