Switch back to the original BigInt library
I am now one of the maintainers of the original BigInt library and will
make sure that it is always supported on the newest Swift version and
also on Linux.

This is not a breaking change for most people as the imports stay the same (the module
name was the same) but may be for some if they explicitely depended on
the fork. Those people should switch back to the original library as I'm
dropping support for the fork in favour of the original one.